requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
pinecone
tiktoken
pypdf
python-dotenv
